+++ lib/include/Apunte.hpp
@@ -1,18 +1,20 @@
 #pragma once
+
 #include <string>
 #include <vector>
+#include <iomanip>
 #include "Comentario.hpp"
 #include "Clase.hpp"
 #include "Usuario.hpp" // Incluir el archivo de encabezado de Comentario
 
-
 class Apunte {
+
 private:
     int ID;
     Clase* clase;
     Usuario* usuario;
     std::string Contenido;
-    std::string Fecha;
+    std::tm Fecha;
     std::vector<Comentario*> Comentarios;
     int Like;
     int Dislike;
@@ -20,8 +22,8 @@ private:
 
 public:
     // Constructor
-    Apunte(int id, Clase* clase, Usuario* usuario, const std::string& contenido,
-           const std::string& fecha, const std::vector<Comentario*>& comentarios,
+    Apunte( Clase* clase, Usuario* usuario, const std::string& contenido,
+           const std::tm& fecha, const std::vector<Comentario*>& comentarios,
            int like, int dislike, int popularidad);
 
     // Destructor
@@ -32,7 +34,7 @@ public:
     Clase* getClase() const;
     Usuario* getUsuario() const;
     std::string getContenido() const;
-    std::string getFecha() const;
+    std::tm getFecha() const;
     std::vector<Comentario*> getComentarios() const;
     int getLike() const;
     int getDislike() const;
@@ -43,10 +45,13 @@ public:
     void setClase(Clase* clase);
     void setUsuario(Usuario* usuario);
     void setContenido(const std::string& contenido);
-    void setFecha(const std::string& fecha);
+    void setFecha(const std::tm& fecha);
     void setComentarios(const std::vector<Comentario*>& comentarios);
     void setLike(int like);
     void setDislike(int dislike);
     void setPopularidad(int popularidad);
+
+    // Método para convertir Apunte en cadena
+    std::string toString();
 };
