+++ lib/src/Periodo.cpp
@@ -1,0 +1,1 @@
+#include <Periodo.hpp>
