--- lib/src/Semestre.cpp
@@ -1,1 +1,0 @@
-#include <Semestre.hpp>
