+++ main/src/Mecha.cpp
@@ -1,0 +1,6 @@
+#include <iostream>
+
+int main(int argc, char* argv[])
+{
+    std::cout << "MECHA APP";
+}
