+++ test/src/Mecha.cpp
@@ -1,5 +1,3 @@
-#pragma once
-
 #include <iostream>
 #include <Apunte.hpp>
 
@@ -67,7 +65,7 @@ int main()
         std::cout << "Testing Agregar Apunte a Clase";
         // Apunte ...
        //  Clase...
-         Apunte::asignar_clase( 1 ); // funciones como esperamos 
+       //  Apunte::asignar_clase( 1 ); // funciones como esperamos 
 
          //  esperamos que la clase tenga ese apunte
